more refactor, rm blocktime test derails
